EID-1111 - Clean up classes and make them consistent
