+++ proxy-node-gateway/src/main/java/uk/gov/ida/notification/GatewayApplication.java
@@ -1,6 +1,5 @@
-package uk.gov.ida.notification.translator;
+package uk.gov.ida.notification;
 
-import com.google.common.collect.ImmutableList;
 import io.dropwizard.Application;
 import io.dropwizard.configuration.EnvironmentVariableSubstitutor;
 import io.dropwizard.configuration.SubstitutingSourceProvider;
@@ -10,29 +9,33 @@ import io.dropwizard.views.ViewBundle;
 import net.shibboleth.utilities.java.support.component.ComponentInitializationException;
 import org.opensaml.core.config.InitializationException;
 import org.opensaml.core.config.InitializationService;
-import org.opensaml.saml.saml2.encryption.Decrypter;
+import org.opensaml.saml.metadata.resolver.MetadataResolver;
 import org.opensaml.saml.security.impl.MetadataCredentialResolver;
 import org.opensaml.security.credential.BasicCredential;
-import org.opensaml.security.credential.Credential;
 import org.opensaml.xmlsec.config.DefaultSecurityConfigurationBootstrap;
 import org.opensaml.xmlsec.keyinfo.KeyInfoCredentialResolver;
 import org.opensaml.xmlsec.signature.support.impl.ExplicitKeySignatureTrustEngine;
-import uk.gov.ida.notification.EidasResponseGenerator;
-import uk.gov.ida.notification.healthcheck.TranslatorHealthCheck;
-import uk.gov.ida.notification.VerifySamlInitializer;
 import uk.gov.ida.notification.exceptions.mappers.AuthnRequestExceptionMapper;
 import uk.gov.ida.notification.exceptions.mappers.HubResponseExceptionMapper;
+import uk.gov.ida.notification.healthcheck.GatewayHealthCheck;
 import uk.gov.ida.notification.pki.KeyPairConfiguration;
+import uk.gov.ida.notification.resources.EidasAuthnRequestResource;
+import uk.gov.ida.notification.resources.HubResponseResource;
+import uk.gov.ida.notification.saml.ResponseAssertionDecrypter;
 import uk.gov.ida.notification.saml.SamlObjectSigner;
 import uk.gov.ida.notification.saml.converters.AuthnRequestParameterProvider;
 import uk.gov.ida.notification.saml.converters.ResponseParameterProvider;
 import uk.gov.ida.notification.saml.metadata.Metadata;
 import uk.gov.ida.notification.saml.metadata.MetadataCredentialResolverInitializer;
-import uk.gov.ida.notification.saml.translation.HubResponseTranslator;
+import uk.gov.ida.notification.saml.EidasAuthnRequestTranslator;
+import uk.gov.ida.notification.saml.validation.EidasAuthnRequestValidator;
 import uk.gov.ida.notification.saml.validation.HubResponseValidator;
 import uk.gov.ida.notification.saml.validation.components.LoaValidator;
+import uk.gov.ida.notification.saml.validation.components.RequestIdWatcher;
+import uk.gov.ida.notification.saml.validation.components.RequestIssuerValidator;
+import uk.gov.ida.notification.saml.validation.components.RequestedAttributesValidator;
 import uk.gov.ida.notification.saml.validation.components.ResponseAttributesValidator;
-import uk.gov.ida.notification.translator.resources.HubResponseFromGatewayResource;
+import uk.gov.ida.notification.saml.validation.components.SpTypeValidator;
 import uk.gov.ida.saml.core.validators.DestinationValidator;
 import uk.gov.ida.saml.core.validators.assertion.AssertionAttributeStatementValidator;
 import uk.gov.ida.saml.core.validators.assertion.AuthnStatementAssertionValidator;
@@ -46,28 +49,29 @@ import uk.gov.ida.saml.hub.transformers.inbound.SamlStatusToIdaStatusCodeMapper;
 import uk.gov.ida.saml.hub.validators.response.idp.IdpResponseValidator;
 import uk.gov.ida.saml.hub.validators.response.idp.components.EncryptedResponseFromIdpValidator;
 import uk.gov.ida.saml.hub.validators.response.idp.components.ResponseAssertionsFromIdpValidator;
+import uk.gov.ida.saml.metadata.MetadataConfiguration;
+import uk.gov.ida.saml.metadata.MetadataHealthCheck;
 import uk.gov.ida.saml.metadata.bundle.MetadataResolverBundle;
-import uk.gov.ida.saml.security.AssertionDecrypter;
-import uk.gov.ida.saml.security.DecrypterFactory;
 import uk.gov.ida.saml.security.MetadataBackedSignatureValidator;
 import uk.gov.ida.saml.security.SamlAssertionsSignatureValidator;
 import uk.gov.ida.saml.security.SamlMessageSignatureValidator;
-import uk.gov.ida.saml.security.validators.encryptedelementtype.EncryptionAlgorithmValidator;
 import uk.gov.ida.saml.security.validators.issuer.IssuerValidator;
 import uk.gov.ida.saml.security.validators.signature.SamlRequestSignatureValidator;
 import uk.gov.ida.saml.security.validators.signature.SamlResponseSignatureValidator;
 
 import java.net.URI;
-import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 
-public class TranslatorApplication extends Application<TranslatorConfiguration> {
+public class GatewayApplication extends Application<GatewayConfiguration> {
 
     private Metadata connectorMetadata;
 
-    private MetadataResolverBundle<TranslatorConfiguration> hubMetadataResolverBundle;
-    private MetadataResolverBundle<TranslatorConfiguration> connectorMetadataResolverBundle;
+    private MetadataResolverBundle<GatewayConfiguration> hubMetadataResolverBundle;
+    private MetadataResolverBundle<GatewayConfiguration> connectorMetadataResolverBundle;
 
+    @SuppressWarnings("WeakerAccess") // Needed for DropwizardAppRules
+    public GatewayApplication() {
+    }
 
     public static void main(final String[] args) throws Exception {
         if (args == null || args.length == 0) {
@@ -77,24 +81,24 @@ public class TranslatorApplication extends Application<TranslatorConfiguration> 
                 throw new RuntimeException("CONFIG_FILE environment variable should be set with path to configuration file");
             }
 
-            new TranslatorApplication().run("server", configFile);
+            new GatewayApplication().run("server", configFile);
         } else {
-            new TranslatorApplication().run(args);
+            new GatewayApplication().run(args);
         }
     }
 
     @Override
     public String getName() {
-        return "translator";
+        return "EidasProxyNode";
     }
 
     @Override
-    public void initialize(final Bootstrap<TranslatorConfiguration> bootstrap) {
+    public void initialize(final Bootstrap<GatewayConfiguration> bootstrap) {
         // Needed to correctly interpolate environment variables in config file
         bootstrap.setConfigurationSourceProvider(
-                new SubstitutingSourceProvider(bootstrap.getConfigurationSourceProvider(),
-                        new EnvironmentVariableSubstitutor(false)
-                )
+            new SubstitutingSourceProvider(bootstrap.getConfigurationSourceProvider(),
+                new EnvironmentVariableSubstitutor(false)
+            )
         );
 
         // Needed to initialise OpenSAML libraries
@@ -113,27 +117,38 @@ public class TranslatorApplication extends Application<TranslatorConfiguration> 
         bootstrap.addBundle(new ViewBundle<>());
 
         // Metadata
-        hubMetadataResolverBundle = new MetadataResolverBundle<>(TranslatorConfiguration::getHubMetadataConfiguration);
+        hubMetadataResolverBundle = new MetadataResolverBundle<>(GatewayConfiguration::getHubMetadataConfiguration);
         bootstrap.addBundle(hubMetadataResolverBundle);
 
-        connectorMetadataResolverBundle = new MetadataResolverBundle<>(TranslatorConfiguration::getConnectorMetadataConfiguration);
+        connectorMetadataResolverBundle = new MetadataResolverBundle<>(GatewayConfiguration::getConnectorMetadataConfiguration);
         bootstrap.addBundle(connectorMetadataResolverBundle);
-
     }
 
     @Override
-    public void run(final TranslatorConfiguration configuration,
-                    final Environment environment) throws ComponentInitializationException {
+    public void run(final GatewayConfiguration configuration,
+                    final Environment environment) throws
+            ComponentInitializationException {
 
         connectorMetadata = createMetadata(connectorMetadataResolverBundle);
+        GatewayHealthCheck gatewayHealthCheck = new GatewayHealthCheck();
 
-        TranslatorHealthCheck translatorHealthCheck = new TranslatorHealthCheck();
-        environment.healthChecks().register(translatorHealthCheck.getName(), translatorHealthCheck);
+        environment.healthChecks().register(gatewayHealthCheck.getName(), gatewayHealthCheck);
+
+        registerMetadataHealthCheck(
+                hubMetadataResolverBundle.getMetadataResolver(),
+                configuration.getHubMetadataConfiguration(),
+                environment,
+                "hub-metadata");
+
+        registerMetadataHealthCheck(
+                connectorMetadataResolverBundle.getMetadataResolver(),
+                configuration.getConnectorMetadataConfiguration(),
+                environment,
+                "connector-metadata");
 
         registerProviders(environment);
         registerExceptionMappers(environment);
         registerResources(configuration, environment);
-
     }
 
     private void registerProviders(Environment environment) {
@@ -146,66 +161,81 @@ public class TranslatorApplication extends Application<TranslatorConfiguration> 
         environment.jersey().register(new AuthnRequestExceptionMapper());
     }
 
-    private void registerResources(TranslatorConfiguration configuration, Environment environment) throws ComponentInitializationException {
-        EidasResponseGenerator eidasResponseGenerator = createEidasResponseGenerator(configuration);
+    private void registerResources(GatewayConfiguration configuration, Environment environment) throws ComponentInitializationException {
+        SamlFormViewBuilder samlFormViewBuilder = new SamlFormViewBuilder();
+        RequestIdWatcher requestIdWatcher = new RequestIdWatcher();
 
+        HubAuthnRequestGenerator hubAuthnRequestGenerator = createHubAuthnRequestGenerator(configuration);
+
+        EidasAuthnRequestValidator eidasAuthnRequestValidator = createEidasAuthnRequestValidator();
         HubResponseValidator hubResponseValidator = createHubResponseValidator(configuration);
 
+        SamlRequestSignatureValidator samlRequestSignatureValidator = createSamlRequestSignatureValidator(connectorMetadataResolverBundle);
 
-        environment.jersey().register(new HubResponseFromGatewayResource(
-                eidasResponseGenerator,
+        environment.jersey().register(new EidasAuthnRequestResource(
+                configuration,
+                hubAuthnRequestGenerator,
+                samlFormViewBuilder,
+                eidasAuthnRequestValidator,
+                samlRequestSignatureValidator,
+                requestIdWatcher));
+
+        environment.jersey().register(new HubResponseResource(
+                samlFormViewBuilder,
                 configuration.getConnectorNodeUrl().toString(),
-                configuration.getConnectorMetadataConfiguration().getExpectedEntityId(),
-                connectorMetadata,
-                hubResponseValidator));
+                hubResponseValidator,
+                environment,
+                configuration.getTranslatorUrl().toString(),
+                requestIdWatcher
+            ));
     }
 
-    private EidasResponseGenerator createEidasResponseGenerator(TranslatorConfiguration configuration) {
-        HubResponseTranslator hubResponseTranslator = new HubResponseTranslator(
-                configuration.getConnectorNodeIssuerId(),
-                configuration.getConnectorNodeUrl().toString(),
-                configuration.getProxyNodeMetadataForConnectorNodeUrl().toString()
+    public void registerMetadataHealthCheck(MetadataResolver metadataResolver, MetadataConfiguration connectorMetadataConfiguration, Environment environment, String name) {
+        MetadataHealthCheck metadataHealthCheck = new MetadataHealthCheck(
+                metadataResolver,
+                name,
+                connectorMetadataConfiguration.getExpectedEntityId()
         );
-        SamlObjectSigner signer = new SamlObjectSigner(
-                configuration.getConnectorFacingSigningKeyPair().getPublicKey().getPublicKey(),
-                configuration.getConnectorFacingSigningKeyPair().getPrivateKey().getPrivateKey(),
-                configuration.getConnectorFacingSigningKeyPair().getPublicKey().getCert()
-        );
-        return new EidasResponseGenerator(hubResponseTranslator, signer);
+
+        environment.healthChecks().register(metadataHealthCheck.getName(), metadataHealthCheck);
     }
 
-    private HubResponseValidator createHubResponseValidator(TranslatorConfiguration configuration) throws ComponentInitializationException {
+    private HubResponseValidator createHubResponseValidator(GatewayConfiguration configuration) throws ComponentInitializationException {
         URI proxyNodeResponseUrl = configuration.getProxyNodeResponseUrl();
         String proxyNodeEntityId = configuration.getProxyNodeEntityId();
 
         SamlMessageSignatureValidator hubResponseMessageSignatureValidator = createSamlMessagesSignatureValidator(hubMetadataResolverBundle);
+        ResponseAssertionDecrypter responseAssertionDecrypter = createDecrypter(configuration.getHubFacingEncryptionKeyPair());
 
         IdpResponseValidator idpResponseValidator = new IdpResponseValidator(
-                new SamlResponseSignatureValidator(hubResponseMessageSignatureValidator),
-                createDecrypter(configuration.getHubFacingEncryptionKeyPair()),
-                new SamlAssertionsSignatureValidator(hubResponseMessageSignatureValidator),
-                new EncryptedResponseFromIdpValidator<>(new SamlStatusToIdaStatusCodeMapper()),
-                new DestinationValidator(proxyNodeResponseUrl, proxyNodeResponseUrl.getPath()),
-                createResponseAssertionsFromIdpValidator(proxyNodeEntityId)
+            new SamlResponseSignatureValidator(hubResponseMessageSignatureValidator),
+                responseAssertionDecrypter.getAssertionDecrypter(),
+            new SamlAssertionsSignatureValidator(hubResponseMessageSignatureValidator),
+            new EncryptedResponseFromIdpValidator<>(new SamlStatusToIdaStatusCodeMapper()),
+            new DestinationValidator(proxyNodeResponseUrl, proxyNodeResponseUrl.getPath()),
+            createResponseAssertionsFromIdpValidator(proxyNodeEntityId)
         );
         ResponseAttributesValidator responseAttributesValidator = new ResponseAttributesValidator();
-        return new HubResponseValidator(idpResponseValidator, responseAttributesValidator, new LoaValidator());
+        return new HubResponseValidator(
+            idpResponseValidator,
+            responseAttributesValidator,
+            new LoaValidator());
     }
 
     private ResponseAssertionsFromIdpValidator createResponseAssertionsFromIdpValidator(String proxyNodeEntityId) {
         IdentityProviderAssertionValidator assertionValidator = new IdentityProviderAssertionValidator(
-                new IssuerValidator(),
-                new AssertionSubjectValidator(),
-                new AssertionAttributeStatementValidator(),
-                new AssertionSubjectConfirmationValidator()
+            new IssuerValidator(),
+            new AssertionSubjectValidator(),
+            new AssertionAttributeStatementValidator(),
+            new AssertionSubjectConfirmationValidator()
         );
         DuplicateAssertionValidator duplicateAssertionValidator = new DuplicateAssertionValidator(new ConcurrentHashMap<>());
         return new ResponseAssertionsFromIdpValidator(
-                assertionValidator,
-                new MatchingDatasetAssertionValidator(duplicateAssertionValidator),
-                new AuthnStatementAssertionValidator(duplicateAssertionValidator),
-                new IPAddressValidator(),
-                proxyNodeEntityId
+            assertionValidator,
+            new MatchingDatasetAssertionValidator(duplicateAssertionValidator),
+            new AuthnStatementAssertionValidator(duplicateAssertionValidator),
+            new IPAddressValidator(),
+            proxyNodeEntityId
         );
     }
 
@@ -214,17 +244,20 @@ public class TranslatorApplication extends Application<TranslatorConfiguration> 
         return new Metadata(metadataCredentialResolver);
     }
 
-    private AssertionDecrypter createDecrypter(KeyPairConfiguration configuration) {
+    private ResponseAssertionDecrypter createDecrypter(KeyPairConfiguration configuration) {
         BasicCredential decryptionCredential = new BasicCredential(
                 configuration.getPublicKey().getPublicKey(),
                 configuration.getPrivateKey().getPrivateKey()
         );
-        List<Credential> decryptionCredentials = ImmutableList.of(decryptionCredential);
-        Decrypter decrypter = new DecrypterFactory().createDecrypter(decryptionCredentials);
-        EncryptionAlgorithmValidator encryptionAlgorithmValidator = new EncryptionAlgorithmValidator();
-        return new AssertionDecrypter(
-                encryptionAlgorithmValidator,
-                decrypter
+        return new ResponseAssertionDecrypter(decryptionCredential);
+    }
+
+    private EidasAuthnRequestValidator createEidasAuthnRequestValidator() {
+        return new EidasAuthnRequestValidator(
+            new RequestIssuerValidator(),
+            new SpTypeValidator(),
+            new LoaValidator(),
+            new RequestedAttributesValidator()
         );
     }
 
@@ -239,5 +272,17 @@ public class TranslatorApplication extends Application<TranslatorConfiguration> 
     private SamlRequestSignatureValidator createSamlRequestSignatureValidator(MetadataResolverBundle hubMetadataResolverBundle) throws ComponentInitializationException {
         SamlMessageSignatureValidator samlMessageSignatureValidator = createSamlMessagesSignatureValidator(hubMetadataResolverBundle);
         return new SamlRequestSignatureValidator(samlMessageSignatureValidator);
+    }
+
+    private HubAuthnRequestGenerator createHubAuthnRequestGenerator(GatewayConfiguration configuration) {
+        EidasAuthnRequestTranslator eidasAuthnRequestTranslator = new EidasAuthnRequestTranslator(
+            configuration.getProxyNodeEntityId(),
+            configuration.getHubUrl().toString());
+        SamlObjectSigner signer = new SamlObjectSigner(
+            configuration.getHubFacingSigningKeyPair().getPublicKey().getPublicKey(),
+            configuration.getHubFacingSigningKeyPair().getPrivateKey().getPrivateKey(),
+            configuration.getHubFacingSigningKeyPair().getPublicKey().getCert()
+        );
+        return new HubAuthnRequestGenerator(eidasAuthnRequestTranslator, signer);
     }
 }
