+++ proxy-node-gateway/src/main/java/uk/gov/ida/notification/HubAuthnRequestGenerator.java
@@ -2,8 +2,8 @@ package uk.gov.ida.notification;
 
 import org.opensaml.saml.saml2.core.AuthnRequest;
 import uk.gov.ida.notification.saml.SamlObjectSigner;
-import uk.gov.ida.notification.saml.translation.EidasAuthnRequest;
-import uk.gov.ida.notification.saml.translation.EidasAuthnRequestTranslator;
+import uk.gov.ida.notification.saml.EidasAuthnRequest;
+import uk.gov.ida.notification.saml.EidasAuthnRequestTranslator;
 
 public class HubAuthnRequestGenerator {
     private final EidasAuthnRequestTranslator translator;
