+++ proxy-node-gateway/src/main/java/uk/gov/ida/notification/saml/HubMdsAssertion.java
@@ -1,12 +1,10 @@
-package uk.gov.ida.notification.saml.translation;
+package uk.gov.ida.notification.saml;
 
 import org.joda.time.DateTime;
 import org.opensaml.saml.saml2.core.Assertion;
 import org.opensaml.saml.saml2.core.Attribute;
 import org.opensaml.saml.saml2.core.AttributeValue;
-import uk.gov.ida.notification.exceptions.hubresponse.HubResponseTranslationException;
 
-import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
