+++ proxy-node-gateway/src/test/java/uk/gov/ida/notification/saml/EidasAuthnRequestTest.java
@@ -1,11 +1,10 @@
-package uk.gov.ida.notification.saml.translation;
+package uk.gov.ida.notification.saml;
 
 import org.junit.Test;
 import org.opensaml.saml.saml2.core.AuthnRequest;
 import se.litsec.eidas.opensaml.common.EidasConstants;
 import uk.gov.ida.notification.SamlInitializedTest;
 import uk.gov.ida.notification.helpers.FileHelpers;
-import uk.gov.ida.notification.saml.SamlParser;
 
 import static org.junit.Assert.assertEquals;
