+++ proxy-node-gateway/src/test/java/uk/gov/ida/notification/saml/EidasAuthnRequestTranslatorTest.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.notification.saml.translation;
+package uk.gov.ida.notification.saml;
 
 import org.junit.Test;
 import org.opensaml.saml.saml2.core.AuthnRequest;
