+++ proxy-node-shared/src/main/java/uk/gov/ida/notification/saml/HubMdsAssertion.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.notification.saml.translation;
+package uk.gov.ida.notification.saml;
 
 import org.joda.time.DateTime;
 import org.opensaml.saml.saml2.core.Assertion;
