+++ proxy-node-translator/build.gradle
@@ -9,7 +9,7 @@ group = 'uk.gov.ida.notification.translator'
 version = "${System.env.BUILD_NUMBER ?: 'SNAPSHOT'}"
 
 apply plugin: 'application'
-mainClassName = 'uk.gov.ida.notification.translator.TranslatorApplication'
+mainClassName = 'uk.gov.ida.notification.TranslatorApplication'
 
 run {
     args 'server', './src/dist/config.yml'
