+++ proxy-node-translator/src/main/java/uk/gov/ida/notification/TranslatorApplication.java
@@ -1,5 +1,6 @@
 package uk.gov.ida.notification;
 
+import com.google.common.collect.ImmutableList;
 import io.dropwizard.Application;
 import io.dropwizard.configuration.EnvironmentVariableSubstitutor;
 import io.dropwizard.configuration.SubstitutingSourceProvider;
@@ -9,33 +10,27 @@ import io.dropwizard.views.ViewBundle;
 import net.shibboleth.utilities.java.support.component.ComponentInitializationException;
 import org.opensaml.core.config.InitializationException;
 import org.opensaml.core.config.InitializationService;
-import org.opensaml.saml.metadata.resolver.MetadataResolver;
+import org.opensaml.saml.saml2.encryption.Decrypter;
 import org.opensaml.saml.security.impl.MetadataCredentialResolver;
 import org.opensaml.security.credential.BasicCredential;
+import org.opensaml.security.credential.Credential;
 import org.opensaml.xmlsec.config.DefaultSecurityConfigurationBootstrap;
 import org.opensaml.xmlsec.keyinfo.KeyInfoCredentialResolver;
 import org.opensaml.xmlsec.signature.support.impl.ExplicitKeySignatureTrustEngine;
+import uk.gov.ida.notification.healthcheck.TranslatorHealthCheck;
 import uk.gov.ida.notification.exceptions.mappers.AuthnRequestExceptionMapper;
 import uk.gov.ida.notification.exceptions.mappers.HubResponseExceptionMapper;
-import uk.gov.ida.notification.healthcheck.GatewayHealthCheck;
 import uk.gov.ida.notification.pki.KeyPairConfiguration;
-import uk.gov.ida.notification.resources.EidasAuthnRequestResource;
-import uk.gov.ida.notification.resources.HubResponseResource;
-import uk.gov.ida.notification.saml.ResponseAssertionDecrypter;
 import uk.gov.ida.notification.saml.SamlObjectSigner;
 import uk.gov.ida.notification.saml.converters.AuthnRequestParameterProvider;
 import uk.gov.ida.notification.saml.converters.ResponseParameterProvider;
 import uk.gov.ida.notification.saml.metadata.Metadata;
 import uk.gov.ida.notification.saml.metadata.MetadataCredentialResolverInitializer;
-import uk.gov.ida.notification.saml.translation.EidasAuthnRequestTranslator;
-import uk.gov.ida.notification.saml.validation.EidasAuthnRequestValidator;
+import uk.gov.ida.notification.saml.HubResponseTranslator;
 import uk.gov.ida.notification.saml.validation.HubResponseValidator;
 import uk.gov.ida.notification.saml.validation.components.LoaValidator;
-import uk.gov.ida.notification.saml.validation.components.RequestIdWatcher;
-import uk.gov.ida.notification.saml.validation.components.RequestIssuerValidator;
-import uk.gov.ida.notification.saml.validation.components.RequestedAttributesValidator;
 import uk.gov.ida.notification.saml.validation.components.ResponseAttributesValidator;
-import uk.gov.ida.notification.saml.validation.components.SpTypeValidator;
+import uk.gov.ida.notification.resources.HubResponseFromGatewayResource;
 import uk.gov.ida.saml.core.validators.DestinationValidator;
 import uk.gov.ida.saml.core.validators.assertion.AssertionAttributeStatementValidator;
 import uk.gov.ida.saml.core.validators.assertion.AuthnStatementAssertionValidator;
@@ -49,29 +44,28 @@ import uk.gov.ida.saml.hub.transformers.inbound.SamlStatusToIdaStatusCodeMapper;
 import uk.gov.ida.saml.hub.validators.response.idp.IdpResponseValidator;
 import uk.gov.ida.saml.hub.validators.response.idp.components.EncryptedResponseFromIdpValidator;
 import uk.gov.ida.saml.hub.validators.response.idp.components.ResponseAssertionsFromIdpValidator;
-import uk.gov.ida.saml.metadata.MetadataConfiguration;
-import uk.gov.ida.saml.metadata.MetadataHealthCheck;
 import uk.gov.ida.saml.metadata.bundle.MetadataResolverBundle;
+import uk.gov.ida.saml.security.AssertionDecrypter;
+import uk.gov.ida.saml.security.DecrypterFactory;
 import uk.gov.ida.saml.security.MetadataBackedSignatureValidator;
 import uk.gov.ida.saml.security.SamlAssertionsSignatureValidator;
 import uk.gov.ida.saml.security.SamlMessageSignatureValidator;
+import uk.gov.ida.saml.security.validators.encryptedelementtype.EncryptionAlgorithmValidator;
 import uk.gov.ida.saml.security.validators.issuer.IssuerValidator;
 import uk.gov.ida.saml.security.validators.signature.SamlRequestSignatureValidator;
 import uk.gov.ida.saml.security.validators.signature.SamlResponseSignatureValidator;
 
 import java.net.URI;
+import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 
-public class EidasProxyNodeApplication extends Application<EidasProxyNodeConfiguration> {
+public class TranslatorApplication extends Application<TranslatorConfiguration> {
 
     private Metadata connectorMetadata;
 
-    private MetadataResolverBundle<EidasProxyNodeConfiguration> hubMetadataResolverBundle;
-    private MetadataResolverBundle<EidasProxyNodeConfiguration> connectorMetadataResolverBundle;
+    private MetadataResolverBundle<TranslatorConfiguration> hubMetadataResolverBundle;
+    private MetadataResolverBundle<TranslatorConfiguration> connectorMetadataResolverBundle;
 
-    @SuppressWarnings("WeakerAccess") // Needed for DropwizardAppRules
-    public EidasProxyNodeApplication() {
-    }
 
     public static void main(final String[] args) throws Exception {
         if (args == null || args.length == 0) {
@@ -81,24 +75,24 @@ public class EidasProxyNodeApplication extends Application<EidasProxyNodeConfigu
                 throw new RuntimeException("CONFIG_FILE environment variable should be set with path to configuration file");
             }
 
-            new EidasProxyNodeApplication().run("server", configFile);
+            new TranslatorApplication().run("server", configFile);
         } else {
-            new EidasProxyNodeApplication().run(args);
+            new TranslatorApplication().run(args);
         }
     }
 
     @Override
     public String getName() {
-        return "EidasProxyNode";
+        return "translator";
     }
 
     @Override
-    public void initialize(final Bootstrap<EidasProxyNodeConfiguration> bootstrap) {
+    public void initialize(final Bootstrap<TranslatorConfiguration> bootstrap) {
         // Needed to correctly interpolate environment variables in config file
         bootstrap.setConfigurationSourceProvider(
-            new SubstitutingSourceProvider(bootstrap.getConfigurationSourceProvider(),
-                new EnvironmentVariableSubstitutor(false)
-            )
+                new SubstitutingSourceProvider(bootstrap.getConfigurationSourceProvider(),
+                        new EnvironmentVariableSubstitutor(false)
+                )
         );
 
         // Needed to initialise OpenSAML libraries
@@ -117,38 +111,27 @@ public class EidasProxyNodeApplication extends Application<EidasProxyNodeConfigu
         bootstrap.addBundle(new ViewBundle<>());
 
         // Metadata
-        hubMetadataResolverBundle = new MetadataResolverBundle<>(EidasProxyNodeConfiguration::getHubMetadataConfiguration);
+        hubMetadataResolverBundle = new MetadataResolverBundle<>(TranslatorConfiguration::getHubMetadataConfiguration);
         bootstrap.addBundle(hubMetadataResolverBundle);
 
-        connectorMetadataResolverBundle = new MetadataResolverBundle<>(EidasProxyNodeConfiguration::getConnectorMetadataConfiguration);
+        connectorMetadataResolverBundle = new MetadataResolverBundle<>(TranslatorConfiguration::getConnectorMetadataConfiguration);
         bootstrap.addBundle(connectorMetadataResolverBundle);
+
     }
 
     @Override
-    public void run(final EidasProxyNodeConfiguration configuration,
-                    final Environment environment) throws
-            ComponentInitializationException {
+    public void run(final TranslatorConfiguration configuration,
+                    final Environment environment) throws ComponentInitializationException {
 
         connectorMetadata = createMetadata(connectorMetadataResolverBundle);
-        GatewayHealthCheck gatewayHealthCheck = new GatewayHealthCheck();
 
-        environment.healthChecks().register(gatewayHealthCheck.getName(), gatewayHealthCheck);
-
-        registerMetadataHealthCheck(
-                hubMetadataResolverBundle.getMetadataResolver(),
-                configuration.getHubMetadataConfiguration(),
-                environment,
-                "hub-metadata");
-
-        registerMetadataHealthCheck(
-                connectorMetadataResolverBundle.getMetadataResolver(),
-                configuration.getConnectorMetadataConfiguration(),
-                environment,
-                "connector-metadata");
+        TranslatorHealthCheck translatorHealthCheck = new TranslatorHealthCheck();
+        environment.healthChecks().register(translatorHealthCheck.getName(), translatorHealthCheck);
 
         registerProviders(environment);
         registerExceptionMappers(environment);
         registerResources(configuration, environment);
+
     }
 
     private void registerProviders(Environment environment) {
@@ -161,81 +144,66 @@ public class EidasProxyNodeApplication extends Application<EidasProxyNodeConfigu
         environment.jersey().register(new AuthnRequestExceptionMapper());
     }
 
-    private void registerResources(EidasProxyNodeConfiguration configuration, Environment environment) throws ComponentInitializationException {
-        SamlFormViewBuilder samlFormViewBuilder = new SamlFormViewBuilder();
-        RequestIdWatcher requestIdWatcher = new RequestIdWatcher();
+    private void registerResources(TranslatorConfiguration configuration, Environment environment) throws ComponentInitializationException {
+        EidasResponseGenerator eidasResponseGenerator = createEidasResponseGenerator(configuration);
 
-        HubAuthnRequestGenerator hubAuthnRequestGenerator = createHubAuthnRequestGenerator(configuration);
-
-        EidasAuthnRequestValidator eidasAuthnRequestValidator = createEidasAuthnRequestValidator();
         HubResponseValidator hubResponseValidator = createHubResponseValidator(configuration);
 
-        SamlRequestSignatureValidator samlRequestSignatureValidator = createSamlRequestSignatureValidator(connectorMetadataResolverBundle);
 
-        environment.jersey().register(new EidasAuthnRequestResource(
-                configuration,
-                hubAuthnRequestGenerator,
-                samlFormViewBuilder,
-                eidasAuthnRequestValidator,
-                samlRequestSignatureValidator,
-                requestIdWatcher));
-
-        environment.jersey().register(new HubResponseResource(
-                samlFormViewBuilder,
+        environment.jersey().register(new HubResponseFromGatewayResource(
+                eidasResponseGenerator,
                 configuration.getConnectorNodeUrl().toString(),
-                hubResponseValidator,
-                environment,
-                configuration.getTranslatorUrl().toString(),
-                requestIdWatcher
-            ));
+                configuration.getConnectorMetadataConfiguration().getExpectedEntityId(),
+                connectorMetadata,
+                hubResponseValidator));
     }
 
-    public void registerMetadataHealthCheck(MetadataResolver metadataResolver, MetadataConfiguration connectorMetadataConfiguration, Environment environment, String name) {
-        MetadataHealthCheck metadataHealthCheck = new MetadataHealthCheck(
-                metadataResolver,
-                name,
-                connectorMetadataConfiguration.getExpectedEntityId()
+    private EidasResponseGenerator createEidasResponseGenerator(TranslatorConfiguration configuration) {
+        HubResponseTranslator hubResponseTranslator = new HubResponseTranslator(
+                configuration.getConnectorNodeIssuerId(),
+                configuration.getConnectorNodeUrl().toString(),
+                configuration.getProxyNodeMetadataForConnectorNodeUrl().toString()
         );
-
-        environment.healthChecks().register(metadataHealthCheck.getName(), metadataHealthCheck);
+        SamlObjectSigner signer = new SamlObjectSigner(
+                configuration.getConnectorFacingSigningKeyPair().getPublicKey().getPublicKey(),
+                configuration.getConnectorFacingSigningKeyPair().getPrivateKey().getPrivateKey(),
+                configuration.getConnectorFacingSigningKeyPair().getPublicKey().getCert()
+        );
+        return new EidasResponseGenerator(hubResponseTranslator, signer);
     }
 
-    private HubResponseValidator createHubResponseValidator(EidasProxyNodeConfiguration configuration) throws ComponentInitializationException {
+    private HubResponseValidator createHubResponseValidator(TranslatorConfiguration configuration) throws ComponentInitializationException {
         URI proxyNodeResponseUrl = configuration.getProxyNodeResponseUrl();
         String proxyNodeEntityId = configuration.getProxyNodeEntityId();
 
         SamlMessageSignatureValidator hubResponseMessageSignatureValidator = createSamlMessagesSignatureValidator(hubMetadataResolverBundle);
-        ResponseAssertionDecrypter responseAssertionDecrypter = createDecrypter(configuration.getHubFacingEncryptionKeyPair());
 
         IdpResponseValidator idpResponseValidator = new IdpResponseValidator(
-            new SamlResponseSignatureValidator(hubResponseMessageSignatureValidator),
-                responseAssertionDecrypter.getAssertionDecrypter(),
-            new SamlAssertionsSignatureValidator(hubResponseMessageSignatureValidator),
-            new EncryptedResponseFromIdpValidator<>(new SamlStatusToIdaStatusCodeMapper()),
-            new DestinationValidator(proxyNodeResponseUrl, proxyNodeResponseUrl.getPath()),
-            createResponseAssertionsFromIdpValidator(proxyNodeEntityId)
+                new SamlResponseSignatureValidator(hubResponseMessageSignatureValidator),
+                createDecrypter(configuration.getHubFacingEncryptionKeyPair()),
+                new SamlAssertionsSignatureValidator(hubResponseMessageSignatureValidator),
+                new EncryptedResponseFromIdpValidator<>(new SamlStatusToIdaStatusCodeMapper()),
+                new DestinationValidator(proxyNodeResponseUrl, proxyNodeResponseUrl.getPath()),
+                createResponseAssertionsFromIdpValidator(proxyNodeEntityId)
         );
         ResponseAttributesValidator responseAttributesValidator = new ResponseAttributesValidator();
-        return new HubResponseValidator(
-            idpResponseValidator,
-            responseAttributesValidator,
-            new LoaValidator());
+        return new HubResponseValidator(idpResponseValidator, responseAttributesValidator, new LoaValidator());
     }
 
     private ResponseAssertionsFromIdpValidator createResponseAssertionsFromIdpValidator(String proxyNodeEntityId) {
         IdentityProviderAssertionValidator assertionValidator = new IdentityProviderAssertionValidator(
-            new IssuerValidator(),
-            new AssertionSubjectValidator(),
-            new AssertionAttributeStatementValidator(),
-            new AssertionSubjectConfirmationValidator()
+                new IssuerValidator(),
+                new AssertionSubjectValidator(),
+                new AssertionAttributeStatementValidator(),
+                new AssertionSubjectConfirmationValidator()
         );
         DuplicateAssertionValidator duplicateAssertionValidator = new DuplicateAssertionValidator(new ConcurrentHashMap<>());
         return new ResponseAssertionsFromIdpValidator(
-            assertionValidator,
-            new MatchingDatasetAssertionValidator(duplicateAssertionValidator),
-            new AuthnStatementAssertionValidator(duplicateAssertionValidator),
-            new IPAddressValidator(),
-            proxyNodeEntityId
+                assertionValidator,
+                new MatchingDatasetAssertionValidator(duplicateAssertionValidator),
+                new AuthnStatementAssertionValidator(duplicateAssertionValidator),
+                new IPAddressValidator(),
+                proxyNodeEntityId
         );
     }
 
@@ -244,20 +212,17 @@ public class EidasProxyNodeApplication extends Application<EidasProxyNodeConfigu
         return new Metadata(metadataCredentialResolver);
     }
 
-    private ResponseAssertionDecrypter createDecrypter(KeyPairConfiguration configuration) {
+    private AssertionDecrypter createDecrypter(KeyPairConfiguration configuration) {
         BasicCredential decryptionCredential = new BasicCredential(
                 configuration.getPublicKey().getPublicKey(),
                 configuration.getPrivateKey().getPrivateKey()
         );
-        return new ResponseAssertionDecrypter(decryptionCredential);
-    }
-
-    private EidasAuthnRequestValidator createEidasAuthnRequestValidator() {
-        return new EidasAuthnRequestValidator(
-            new RequestIssuerValidator(),
-            new SpTypeValidator(),
-            new LoaValidator(),
-            new RequestedAttributesValidator()
+        List<Credential> decryptionCredentials = ImmutableList.of(decryptionCredential);
+        Decrypter decrypter = new DecrypterFactory().createDecrypter(decryptionCredentials);
+        EncryptionAlgorithmValidator encryptionAlgorithmValidator = new EncryptionAlgorithmValidator();
+        return new AssertionDecrypter(
+                encryptionAlgorithmValidator,
+                decrypter
         );
     }
 
@@ -272,17 +237,5 @@ public class EidasProxyNodeApplication extends Application<EidasProxyNodeConfigu
     private SamlRequestSignatureValidator createSamlRequestSignatureValidator(MetadataResolverBundle hubMetadataResolverBundle) throws ComponentInitializationException {
         SamlMessageSignatureValidator samlMessageSignatureValidator = createSamlMessagesSignatureValidator(hubMetadataResolverBundle);
         return new SamlRequestSignatureValidator(samlMessageSignatureValidator);
-    }
-
-    private HubAuthnRequestGenerator createHubAuthnRequestGenerator(EidasProxyNodeConfiguration configuration) {
-        EidasAuthnRequestTranslator eidasAuthnRequestTranslator = new EidasAuthnRequestTranslator(
-            configuration.getProxyNodeEntityId(),
-            configuration.getHubUrl().toString());
-        SamlObjectSigner signer = new SamlObjectSigner(
-            configuration.getHubFacingSigningKeyPair().getPublicKey().getPublicKey(),
-            configuration.getHubFacingSigningKeyPair().getPrivateKey().getPrivateKey(),
-            configuration.getHubFacingSigningKeyPair().getPublicKey().getCert()
-        );
-        return new HubAuthnRequestGenerator(eidasAuthnRequestTranslator, signer);
     }
 }
