+++ proxy-node-translator/src/main/java/uk/gov/ida/notification/saml/HubResponseTranslator.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.notification.saml.translation;
+package uk.gov.ida.notification.saml;
 
 import org.opensaml.saml.saml2.core.Attribute;
 import org.opensaml.saml.saml2.core.Response;
